fixed validation

+++ segan_biometrics.py
@@ -67,7 +67,7 @@ def train_model(model, criterion, optimizer,scheduler, num_epochs=25):
 			with torch.set_grad_enabled(True):
 				outputs = model(local_batch)
 				_, preds = torch.max(outputs, 1)
-				local_labels=torch.tensor(to_categorical(local_labels, 27), device=device)
+				local_labels=torch.tensor(to_categorical(local_labels, 27), device=device).float()
 				loss = criterion(outputs, local_labels)
 				# backward + optimize only if in training phase
 				loss.backward()
